backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0

mercantile
datashader
pandas
pillow
matplotlib